refactor(maia-engines): consolidate utils and modules
- Inline attribute-sanitizer, cobinary-preview, event-debounce, markdown into view.engine.js
- Inline upload-progress, deriveInboxRef into actor.engine.js and browser.js
- Merge security-constants + payload-sanitizer into utils/security.js
- Merge trace + perf into utils/debug.js (drop perfUpload)
- Remove dead exports from utils.js (isContentEditableActive, isInsideRoot)
- Slim registry.js: remove clear(), listModuleConfigs(), flatten createModule
- Delete 10 util files, add 2 consolidated (debug.js, security.js)
- 14 util files → 6, ~49% reduction in utils+modules lines
