--- libs/maia-engines/src/utils/security-constants.js
@@ -1,120 +1,0 @@
-/**
- * Centralized security constants for DOM/CSS/expression sanitization.
- * Single source of truth for XSS prevention across ViewEngine, StyleEngine, Evaluator.
- */
-
-/** SECURITY: Block prototype chain / constructor access to prevent exploitation */
-export const FORBIDDEN_PATH_KEYS = ['__proto__', 'constructor', 'prototype']
-
-/** SECURITY: Block CSS injection via url(), expression(), -moz-binding, etc. */
-export const CSS_INJECTION_PATTERNS = [
-	/javascript\s*:/i,
-	/vbscript\s*:/i,
-	/data\s*:\s*[^,]*base64\s*,/i,
-	/expression\s*\(/i,
-	/-moz-binding\s*:/i,
-	/@import\b/i,
-	/behavior\s*:/i,
-]
-
-/** SECURITY: Tag allowlist for createElement - prevents script/iframe/object/embed injection */
-export const SAFE_TAGS = new Set([
-	'div',
-	'span',
-	'p',
-	'a',
-	'button',
-	'input',
-	'textarea',
-	'select',
-	'option',
-	'optgroup',
-	'form',
-	'label',
-	'fieldset',
-	'legend',
-	'img',
-	'picture',
-	'source',
-	'ul',
-	'ol',
-	'li',
-	'dl',
-	'dt',
-	'dd',
-	'table',
-	'thead',
-	'tbody',
-	'tfoot',
-	'tr',
-	'th',
-	'td',
-	'caption',
-	'colgroup',
-	'col',
-	'h1',
-	'h2',
-	'h3',
-	'h4',
-	'h5',
-	'h6',
-	'header',
-	'footer',
-	'main',
-	'nav',
-	'section',
-	'article',
-	'aside',
-	'details',
-	'summary',
-	'figure',
-	'figcaption',
-	'blockquote',
-	'pre',
-	'code',
-	'em',
-	'strong',
-	'small',
-	'sub',
-	'sup',
-	'mark',
-	'del',
-	'ins',
-	'abbr',
-	'time',
-	'progress',
-	'meter',
-	'output',
-	'dialog',
-	'hr',
-	'br',
-	'video',
-	'audio',
-	'canvas',
-	'svg',
-	'path',
-	'circle',
-	'rect',
-	'line',
-	'polyline',
-	'polygon',
-	'text',
-	'g',
-	'defs',
-	'use',
-	'symbol',
-])
-
-/** Boolean HTML attributes (set via element[name] = bool) */
-export const BOOLEAN_ATTRS = new Set([
-	'disabled',
-	'readonly',
-	'checked',
-	'selected',
-	'autofocus',
-	'required',
-	'multiple',
-])
-
-/** SECURITY: URL-bearing attributes - require safe protocol (blocks javascript:, data:, etc.) */
-export const URL_ATTRS = new Set(['href', 'src', 'action', 'formaction', 'poster'])
